TaskHandle: cancel nested coroutines
When outermost (parent) task is canceled, this information will now
be propagated to all its children so that they won't continue either.

That means any external (non-TaskHandle) awaiters that a TaskHandle
coro is coawaiting on, can no longer assume that their await_resume
will always be called when the coro resumes.